code/central/requirements.txt
aliot-py >=1.0.0
pyserial >= 3.5
bleak